Change the upgrade version number from 1.3.0 to 2.0.0
