+++ pkg/bootstrap/upgrade.go
@@ -19,7 +19,7 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/bootstrap/versions/v1_2_1"
 	"github.com/matrixorigin/matrixone/pkg/bootstrap/versions/v1_2_2"
 	"github.com/matrixorigin/matrixone/pkg/bootstrap/versions/v1_2_3"
-	"github.com/matrixorigin/matrixone/pkg/bootstrap/versions/v1_3_0"
+	"github.com/matrixorigin/matrixone/pkg/bootstrap/versions/v2_0_0"
 )
 
 // initUpgrade all versions need create a upgrade handle in pkg/bootstrap/versions
@@ -31,7 +31,7 @@ func (s *service) initUpgrade() {
 	// TODO: When v1.2.1 release, open the commented code as follows, Enable v1.2.2 upgrade package
 	s.handles = append(s.handles, v1_2_2.Handler)
 	s.handles = append(s.handles, v1_2_3.Handler)
-	s.handles = append(s.handles, v1_3_0.Handler)
+	s.handles = append(s.handles, v2_0_0.Handler)
 }
 
 func (s *service) getFinalVersionHandle() VersionHandle {
