+++ pkg/bootstrap/versions/v2_0_0/cluster_upgrade_list.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v1_3_0
+package v2_0_0
 
 import (
 	"fmt"
