+++ pkg/bootstrap/versions/v2_0_0/data_key.go
@@ -12,13 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v1_3_0
+package v2_0_0
 
 import (
-	"github.com/matrixorigin/matrixone/pkg/common/log"
-	"github.com/matrixorigin/matrixone/pkg/common/runtime"
+	"github.com/matrixorigin/matrixone/pkg/cdc"
+	"github.com/matrixorigin/matrixone/pkg/util/executor"
 )
 
-func getLogger(sid string) *log.MOLogger {
-	return runtime.ServiceRuntime(sid).Logger()
+var InsertInitDataKey = func(txn executor.TxnExecutor, kek string) (err error) {
+	sql, err := cdc.GetInitDataKeySql(kek)
+	if err != nil {
+		return
+	}
+	_, err = txn.Exec(sql, executor.StatementOption{})
+	return
 }
