+++ pkg/bootstrap/versions/v2_0_0/upgrade.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v1_3_0
+package v2_0_0
 
 import (
 	"context"
@@ -29,7 +29,7 @@ import (
 var (
 	Handler = &versionHandle{
 		metadata: versions.Version{
-			Version:           "1.3.0",
+			Version:           "2.0.0",
 			MinUpgradeVersion: "1.2.3",
 			UpgradeCluster:    versions.Yes,
 			UpgradeTenant:     versions.Yes,
